feat(kbar): embed api-key/webhook/dataset/inbox/meeting/mail-view create forms in palette
Extract reusable form components from their dialogs (api-key, webhook,
dataset, inbox, meeting, mail-view) so both the existing dialogs and new
command-palette create pages share one form. Add launcher actions and
create-group page for these resources.
