+++ apps/web/src/components/fields/displays/display-rich-text.tsx
@@ -12,6 +12,7 @@ export function DisplayRichText() {
   return (
     <DisplayWrapper copyValue={richText || null}>
       <div
+        data-slot='field-display-rich-text'
         className='prose max-w-none prose-sm dark:prose-invert prose-p:text-sm prose-p:leading-6 prose-p:mb-2 prose-p:mt-0 prose-p:text-muted-foreground prose-a:text-blue-600 prose-a:no-underline hover:prose-a:underline prose-img:rounded-lg prose-img:max-w-full'
         dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(richText) }}
       />
